+++ test/example/index.js
@@ -5,10 +5,6 @@ import { Button } from 'antd'
 
 import style from './index.css'
 
-import five from 'five'
-
-console.log(five())
-
 staff()
 
 console.log('react')